% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{api_scrape}
\alias{api_scrape}
\title{Scrape stats.nba.com API and return statsnbaR data}
\usage{
api_scrape(endpoint, filters)
}
\value{
A data frame or a list of data frames depending on the number of
result sets returned by the JSON API.
}
\description{
Scrape stats.nba.com API and return statsnbaR data
}

