% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{filter_per_player}
\alias{filter_per_player}
\title{Generate filters for a \sQuote{per player} data query.}
\usage{
filter_per_player(...)
}
\arguments{
\item{...}{List of key-value pairs to use as filters in a query to the
per-player data endpoint of stats.nba.com}
}
\value{
A list of key-value pairs for passing to the per-player query
  functions
}
\details{
\describe{
    \item{college}{}
    \item{conference}{}
    \item{country}{}
    \item{date_from}{}
    \item{date_to}{}
    \item{division}{}
    \item{draft_pick}{}
    \item{draft_year}{}
    \item{game_scope}{}
    \item{game_segment}{}
    \item{height_segment}{}
    \item{last_n}{}
    \item{league}{}
    \item{location}{}
    \item{month}{}
    \item{opponent_team}{}
    \item{win_loss}{}
    \item{playoff_round}{}
    \item{pace_adjust}{}
    \item{per}{}
    \item{period}{}
    \item{experience}{}
    \item{position}{}
    \item{plus_minus}{}
    \item{rank}{}
    \item{season}{}
    \item{season_segment}{}
    \item{season_type}{}
    \item{shot_clock}{}
    \item{starting_status}{}
    \item{team_id}{}
    \item{opponent_conference}{}
    \item{opponent_division}{}
    \item{weight_segment}{}
}
}
\seealso{
\code{\link{per_player_base}} \code{\link{per_player_advanced}}
  \code{\link{per_player_miscellaneous}} \code{\link{per_player_scoring}}
  \code{\link{per_player_usage}}
}

