% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{filter_per_player_clutch}
\alias{filter_per_player_clutch}
\title{Generate filters for a \sQuote{per player} data query.}
\usage{
filter_per_player_clutch(...)
}
\arguments{
\item{...}{List of key-value pairs to use as filters in a query to the
per-player data endpoint of stats.nba.com. The additional clutch-related
key-value pairs are described (with their default values indicated at the
end in brackets.
\describe{
  \item{clutch_time}{Number of seconds left in game: 300, 240, ..., 60, 30, 10 (300)}
  \item{lead}{Whether the team is ahead or behind (incl. tied): 'any', 'ahead', 'behind' ('any')}
  \item{point_diff}{The size of the point differential: 1--5 (5)}
}}
}
\value{
A list of key-value pairs for passing to the per-player query
  functions
}
\description{
Generates a list of filters with default values for per player data queries.
}
\details{
This function is the same as the \code{\link{per_player_filters}} function
with the addition of the filters to determine what \sQuote{clutch} time is.
}
\seealso{
\code{\link{per_player_filters}}
}

