% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_player.R
\name{filter_per_player_worker}
\alias{filter_per_player_worker}
\title{Worker function to generate filters for a \sQuote{per player} data query.}
\usage{
filter_per_player_worker(endpoint, ...)
}
\arguments{
\item{endpoint}{List containing stats.nba.com API endpoint specification as
given by statsnbaR YAML data.}

\item{...}{List of key-value pairs to use as the filters argument of
\code{\link{per_player_data}}}
}
\value{
A list of key-value pairs for passing to the per-player query
  functions
}
\description{
\code{\link{filter_per_player}} and \code{\link{filter_per_player_clutch}}
are both essentially wrappers to this function which does the work.
}
\details{
Returns a named list of key-value pairs which are given either by the
arguments to the function, or by default values specified by statnbaR
internal YAML data. Ensures that all the filters specified have a value,
but performs no checks on the values themselves.
}
\author{
Stephen Wade
}
\seealso{
\code{\link{per_player_data}} \code{\link{filter_per_player}} 
  \code{\link{filter_per_player_clutch}}
}
\keyword{internal}

