% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{get_bio}
\alias{get_bio}
\title{Get player biography data}
\usage{
get_bio(filters = filter_bio())
}
\arguments{
\item{filters}{A named list of key-value filters constructed by 
\code{\link{filter_bio}}. Full list of avaiable filters that statsnbaR
recognises is given in the documentation the filter constructor.}
}
\value{
A data.frame containing the player biography data retrieved, with
  columns converted to the data types specified by statsnbaR's internal
  YAML.
}
\description{
Retrieves the player biography data from stats.nba.com subject to the
filters recognised by statsnbaR, and returns it as a data.frame.
}
\details{
TODO : add details
}

