% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_filter_values}
\alias{map_filter_values}
\title{Translate values (only) to stats.nba.com API values}
\usage{
map_filter_values(filters)
}
\arguments{
\item{filters}{The statsnbaR filter key-value pairs}
}
\value{
Named list of key-values with values mapped to stats.nba.com API
values
}
\description{
Translates the values in a named list of statsnbaR key-value pairs to
values that the stats.nba.com API recognises.
}

