% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_filters}
\alias{map_filters}
\title{Translate key-value pairs to the stats.nba.com API key-value pairs}
\usage{
map_filters(filters, api.endpoint)
}
\arguments{
\item{filters}{The statsnbaR filter key-value pairs.}
}
\value{
Named list of key-values mapped to stats.nba.com API key-value
pairs.
}
\description{
Translates the key-values in a named list of statsnbaR key-value pairs to
key-value pairs that the stats.nba.com API recognises.
}

