% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_results}
\alias{map_results}
\title{Map flattened NBA API data to statsnbaR data}
\usage{
map_results(df, api.mapping)
}
\arguments{
\item{df}{Flattened NBA API data}

\item{api.mapping}{The mapping from API data to statsnbaR data}
}
\value{
The data frame with mapped values
}
\description{
Here's a description.
}

