% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_json_headers}
\alias{parse_json_headers}
\title{Parse the returned JSON header information}
\usage{
parse_json_headers(headers)
}
\arguments{
\item{headers}{List contained the header info as returned by stats.nba.com}
}
\value{
character values of column names, stitched together when
}
\description{
This is a dogs breakfast of a hack. It seems that stats.nba.com has some
edge-case header and resultSet behaviour that means we need to stitch
together repeated column names, e.g. fga less than 5ft, fgm less than 5ft,
fg_pct less than 5ft. (and then for 5-9ft and so on).
}
\keyword{internal}

