% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{per_player_advanced}
\alias{per_player_advanced}
\title{Retrieve \sQuote{per player} advanced data}
\usage{
per_player_advanced(filters = per_player_filters(), clutch = FALSE)
}
\description{
Retrieve \sQuote{per player} advanced data
}

