% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{per_player_base}
\alias{per_player_base}
\title{Retrieve \sQuote{per player} base data}
\usage{
per_player_base(filters = per_player_filters(), clutch = FALSE)
}
\description{
Retrieve \sQuote{per player} base data
}

