% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{per_player_data}
\alias{per_player_data}
\title{Retrieve \sQuote{per player} data}
\usage{
per_player_data(filters = filter_per_player(), clutch = FALSE,
  measurement = "base")
}
\value{
A data.frame with the
}
\description{
Retrieve the aggregated data for each player from stats.nba.com
}

