% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{per_player_miscellanous}
\alias{per_player_miscellanous}
\title{Retrieve \sQuote{per player} miscellaneous data}
\usage{
per_player_miscellanous(filters = per_player_filters(), clutch = FALSE)
}
\description{
Retrieve \sQuote{per player} miscellaneous data
}

