% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{per_player_scoring}
\alias{per_player_scoring}
\title{Retrieve \sQuote{per player} scoring data}
\usage{
per_player_scoring(filters = per_player_filters(), clutch = FALSE)
}
\description{
Retrieve \sQuote{per player} scoring data
}

