% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{per_player_usage}
\alias{per_player_usage}
\title{Retrieve \sQuote{per player} usage data}
\usage{
per_player_usage(filters = per_player_filters(), clutch = FALSE)
}
\description{
Retrieve \sQuote{per player} usage data
}

