% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsnbaR-package.R
\docType{package}
\name{statsnbaR-package}
\alias{statsnbaR}
\alias{statsnbaR-package}
\title{R interface to stats.nba.com}
\description{
A simple interface to stats.nba.com providing basic query functionality
}
\details{
This is a simple interface to stats.nba.com based off of the Python package
\href{https://github.com/seemethere/nba_py}{nba_py}.

Before going into any further details of the package, there are some
house-keeping tasks:

All the data from the website is Copyright (c) 2016 NBA Media Ventures, LLC.
All rights reserved. When using this package you must agree to the
\href{http://www.nba.com/news/termsofuse.html}{Terms of Use} of the website.

As this package sends http requests to stats.nba.com, you should read their
their \href{http://www.nba.com/news/privacy_policy.html}{Privacy Policy}
before using this package.

This code is licensed under the MIT license
\url{https://www.r-project.org/Licenses/MIT}, and you may use this package
strictly under those terms.

The details of the API end-points were manually sourced by the approach
given on this blog post:
\url{http://www.gregreda.com/2015/02/15/web-scraping-finding-the-api}.
In order to semi-future-proof this, the queries to these end-points and
the data extracted from them are evaluated through a fairly informal
abstract data layer (ADL). The ADL is specified in internal data extracted,
from a YAML which can be viewed on github:
\url{http://www.github.com/stephematician/statsnbaR/data-raw}.

The package is split into player, team and game data.
}
\author{
Stephen Wade \email{stephematician@gmail.com}
}
\seealso{
\pkg{yaml}, \pkg{httr}
}

