% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{team_game_logs}
\alias{team_game_logs}
\title{Retrieve team game logs}
\usage{
team_game_logs(league, season, season_type)
}
\arguments{
\item{league}{TODO}

\item{season}{TODO}

\item{season_type}{TODO}
}
\value{
TODO
}
\description{
Retrieves all team game logs from stats.nba.com for a given season and
league.
}
\details{
TODO
}
\seealso{
\code{\link{player_game_logs}}
}

