% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_filters}
\alias{valid_filters}
\title{Validate key-value pairs as being statsnbaR-recognisable}
\usage{
valid_filters(filters, allow_na = TRUE)
}
\arguments{
\item{filters}{Named list of key-value pairs used internally by statsnbaR}
}
\value{
Logical result of tests
}
\description{
Validate key-value pairs as being statsnbaR-recognisable
}
\keyword{internal}

