% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_results}
\alias{valid_results}
\title{Check output of API matches basic format}
\usage{
valid_results(api_result)
}
\arguments{
\item{api_result}{The list returned by \code{content} applied to the
result of a query of stats.nba.com}
}
\value{
Logical result of tests
}
\description{
Check output of API matches basic format
}
\keyword{internal}

