% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_resultSet}
\alias{valid_resultSet}
\title{Check output of API matches basic format}
\usage{
valid_resultSet(rs)
}
\arguments{
\item{rs}{a resultSet as they appear in JSON data from stats.nba.com}
}
\value{
logical result of tests
}
\description{
Check output of API matches basic format
}
\keyword{internal}

