% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_player.R
\name{filter_per_player_clutch}
\alias{filter_per_player_clutch}
\title{Generate filters for a \sQuote{per player} data query.}
\usage{
filter_per_player_clutch(...)
}
\arguments{
\item{...}{List of key-value pairs to use as filters in a query to the
per-player data endpoint of stats.nba.com. The basix key-values are
described in \code{\link{filter_per_player}}. The additional clutch-related
key-value pairs are described here with default values indicated in
brackets:
\describe{
  \item{clutch_time}{Number of seconds left in game: 300, 240, ..., 60,
    30, 10 (300)}
  \item{lead}{Whether the team is ahead or behind (incl. tied): 'any',
    'ahead', 'behind' ('any')}
  \item{point_diff}{The size of the point differential: 1--5 (5)}
}}
}
\value{
A list of key-value pairs for passing to the per-player data query
  function \code{\link{per_player_data}} with argument \code{clutch=TRUE}.
}
\description{
Creates a list of key-value pairs from the supplied arguments for use in
\code{\link{per_player_data}} with argument \code{clutch=TRUE}; and inserts
default values for any filters not specified in the call to the function.
}
\details{
This function is the same as the \code{\link{filter_per_player}} function
with the addition of the filters to determine what \sQuote{clutch} time is.
}
\examples{
\dontrun{
  fppc <- filter_per_player_clutch(league='NBA', clutch_time=300)
  ppd <- per_player_data(filters=fppc, clutch=TRUE)
}

}
\seealso{
\code{\link{per_player_data}} \code{\link{filter_per_player}}
}

