% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{filter_per_player_worker}
\alias{filter_per_player_worker}
\title{Generate filters for a \sQuote{per player} data query.}
\usage{
filter_per_player_worker(endpoint, ...)
}
\arguments{
\item{endpoint}{List containing stats.nba.com API endpoing specification.}

\item{...}{List of key-value pairs to use as filters in a query to the
per-player data endpoint of stats.nba.com}
}
\value{
A list of key-value pairs for passing to the per-player query
  functions
}
\description{
Generate filters for a \sQuote{per player} data query.
}
\author{
Stephen Wade
}
\keyword{internal}

