% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_json_rowSet}
\alias{flatten_json_rowSet}
\title{Flatten JSON result sets}
\usage{
flatten_json_rowSet(json_result)
}
\arguments{
\item{json_result}{A list returned by JSON containing the rowSet and headers}
}
\value{
A data.frame for the flattened JSON rowSet
}
\description{
Takes a list provided by JSON and flattens the rowSet element to a
data.frame. This is acheived by forcing any NULL values to NA, and then
applying the unlist function. The data.frame is then returned with the
attribute names set to the headers specified in the JSON data.
}

