% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{get_bio}
\alias{get_bio}
\title{Get player biography data}
\usage{
get_bio(filters = filter_bio(), method = NULL, ...)
}
\arguments{
\item{filters}{A named list of key-value filters constructed by 
\code{\link{filter_bio}}. Full list of avaiable filters that statsnbaR
recognises is given in the documentation the filter constructor.}

\item{method}{Optional user-supplied function to retrieve JSON from 
stats.nba.com}
}
\value{
A data.frame containing the player biography data with
  columns converted to the data types specified by statsnbaR's internal
  YAML.
  \describe{
     \item{person_id}{integer - ID of the player, used in many other
       datasets}
     \item{player_name}{character - the player's first and last name}
     \item{team_id}{integer - the \code{team_id} of the team the player plays for
       according to the season selected (see \code{\link{filter_bio}})}
     \item{team_abbr}{character - abbreviated team name}
     \item{age}{integer - age of player in the season selected (see
       \code{\link{filter_bio}})}
     \item{height}{integer - listed player height in inches}
     \item{weight}{integer - listed player weight in lbs}
     \item{college}{character - name of college or prior affiliation}
     \item{country}{character - name of country of origin of player}
     \item{draft_year}{integer - year player was drafted  (NA if undrafted)}
     \item{draft_round}{integer - round player was drafted in  (NA if undrafted)}
     \item{draft_number}{integer - number player drafted at (NA if undrafted)}
     \item{games}{numeric - games played within the selected filters}
  }
All other returned columns are described in the glossary provided at
\url{http://stats.nba.com/help/glossary}.
}
\description{
Retrieves the player biography data from stats.nba.com subject to the
filters recognised by statsnbaR, and returns it as a data.frame.
}
\details{
Provides player data given as per-season measurements. For a given league and
season, this will return basic and advanced statistics for all the active
players.

The query filters are constructed by the \code{\link{filter_bio}} function.
The list of filters available is given in the documentation for that
constructor. The units that the values/data is aggregated into is determined
by the \code{per} filter, which can be set to either \code{total} for the
total values or \code{game} for per-game values. This won't affect the
percentage statistics returned.

The filters may include season segments, such as the playoffs,
date filters or versus opponents and so on, and the values returned will
reflect the performance and data collected in only those selected periods
or whatever filter is applied.
}
\examples{
\dontrun{
    df <- get_bio(filter_bio(season=2014))
    tail(df)
}

}
\seealso{
\url{http://stats.nba.com/help/glossary/}
}

