% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{get_players}
\alias{get_players}
\title{Retrieve current and historical player data}
\usage{
get_players(league, season)
}
\arguments{
\item{league}{Character value of league (e.g. 'NBA')}

\item{season}{Numeric value of the base season (e.g. 2015 for the 2015-2016
season)}
}
\description{
Retrieves the player data from stats.nba.com
}

