% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{get_players}
\alias{get_players}
\title{Retrieve current and historical player data}
\usage{
get_players(league, season)
}
\arguments{
\item{league}{A character value of league, current recognised values are:
\describe{
  \item{'nba'}{NBA player info}
  \item{'d-league}{D-League player info}
}}

\item{season}{Numeric value of the base season, e.g. 2015 for the 2015-2016
season and so on.}
}
\value{
A data.frame with names of current and historical players and the
  fields
  \describe{
     \item{person_id}{integer - ID of the player, used in many other
       datasets}
     \item{first_name}{character - the player's first name}
     \item{last_name}{character - the player's last name}
     \item{roster_status}{logical - currently rostered in the selected
       season}
     \item{year_start}{integer - year player entered NBA}
     \item{year_end}{integer - year player exited NBA}
     \item{team_id}{integer - ID of the player's team in the season (if 
       roster_status is true)}
     \item{team_city}{character - city of the player's team in the season (if
       roster_status is true)}
     \item{team_name}{character - name of the player's team in the season
       (if roster_status is true)}
     \item{team_abbr}{character - abbrev. name of player's team (if
       roster_status is true)}
     \item{has_played}{logical - did the player record at least one game in
       NBA at any time (even after the selected \code{season})}
  }
}
\description{
Retrieves current and historical player data from stats.nba.com
}
\details{
Get the data for all the players that stats.nba.com recognises for a given
\code{league}, and indicate their status for a given \code{season}.
}
\examples{
\dontrun{
    df <- get_players(season=2014, league='d-league')
    tail(df)
}

}

