% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_filter_values}
\alias{map_filter_values}
\title{Translate values (only) to stats.nba.com API values}
\usage{
map_filter_values(filters)
}
\arguments{
\item{filters}{The statsnbaR filter key-value pairs}
}
\value{
Named list of key-values with \emph{values} mapped to stats.nba.com 
  API values
}
\description{
Translates the \emph{values} in a named list of statsnbaR key-value pairs to
values that the stats.nba.com API recognises.
}
\details{
The mappings for each filter are specified in the internal YAML data
which look, for example, like the following:
\preformatted{
filters :
    league :
        mapping : \{
            NBA      : "00",
            D-League : "20"
            \}
        default : NBA
}
When we say that only the \emph{value} is mapped, that means we do not
map the key-name \sQuote{league} to the key-name that stats.nba.com
expects. We simply map the value 'NBA' that statsnbaR accepts to the value
"00" that the stats.nba.com API expects. Simple, yes?
}
\seealso{
\code{\link{map_filters}}
}
\keyword{internal}

