% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_results}
\alias{map_results}
\title{Map flattened NBA API data to statsnbaR data}
\usage{
map_results(df, api.mapping)
}
\arguments{
\item{df}{Flattened NBA API data}

\item{api.mapping}{The mapping from API data to statsnbaR data}
}
\value{
The data frame with mapped values
}
\description{
This takes the data.frames returned by \code{\link{flatten_json_rowSet}}
and converts the attributes (columns) to the desired data types specified
by statsnbaR's internal YAML.
}
\details{
The column name is used to identify the correct mapping and/or type
conversion function. The mappings and data types are specified in the
statsnbaR internal YAML data, available at
\url{https://www.github.com/stephematician/statsnbaR/tree/master/data-raw/ADL.yaml}

The YAML specification of the data type might look as follows:
\preformatted{
data    :
    played :
        class   : logical
        mapping : \{
            'Y' : TRUE,    # played
            'N' : FALSE    # never played
            \}
}

Now any column named \sQuote{played} would be mapped from the character
values to a logical TRUE or FALSE, and converted to the logical class
(a redundant step for this example).
}
\seealso{
\code{\link{type_converters}} \code{\link{map_result_values}}
}
\keyword{internal}

