% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{per_player_agg}
\alias{per_player_agg}
\title{Retrieve \sQuote{per player} aggregated data}
\usage{
per_player_agg(filters = filter_per_player(), clutch = FALSE,
  measurement = "base")
}
\arguments{
\item{filters}{A named list of key-value filters constructed by either
\code{\link{filter_per_player}} if \code{clutch=FALSE} or
\code{\link{filter_per_player_clutch}} if \code{clutch=TRUE}. Full list of
avaiable filters that statsnbaR recognises is given in the documentation
of those two filter constructor functions.}

\item{clutch}{A logical value indicating whether to extract the
\sQuote{clutch} data, i.e. to only include plays that occurred during
\sQuote{clutch} time in the aggregation.}

\item{measurement}{A character string representing the desired dataset
  \describe{
    \item{base}{The traditional player statistics such as field goals,
      rebounds, assists etc.}
    \item{advanced}{The advanced player statistics such as offensive and
      defensive ratings, effective field goal %age, PIE and pace, and
      others.}
    \item{miscellaneous}{Miscellaneous player statistics such as points in
      the paint, fastbreak points, opponent points in the paint, blocks,
      fouls, etc.}
    \item{scoring}{Scoring player statistics such as the percentage
      of points off turnovers, in the paint, points that were assisted on
      etc. by that player.}
    \item{usage}{Usage statistics which are roughly the traditional
      statistics but measured as the players percentage contribution of the
      teams values while they are on court.}
}}
}
\value{
A data.frame containing the aggregated player data retrieved with
  columns converted to the data types specified by statsnbaR's internal
  YAML.
\describe{
     \item{person_id}{integer - ID of the player, used in many other
       datasets}
     \item{player_name}{character - the player's first and last name}
     \item{team_id}{integer - the \code{team_id} of the team the player plays for
       according to the season selected (see \code{\link{filter_bio}})}
     \item{team_abbr}{character - abbreviated team name}
     \item{age}{integer - age of player in the season selected}
     \item{games}{numeric - games played within the selected filters}
     \item{wins}{numeric - wins in the player's games within the
       selected filters}
     \item{wins}{numeric - losses in the player's games within the
       selected filters}
     \item{mins}{numeric - minutes player was on court in game}
}
All other returned columns are described in the glossary provided at
\url{http://stats.nba.com/help/glossary}.
}
\description{
Retrieve the aggregated data for each player from stats.nba.com
}
\details{
Collects player 'play' statistics as averages or other aggregates over games
in different units such as total values, per game values, per possession
values and so forth.

The statistics are grouped into various categories which are determined
by the value of the \code{measurement} argument. The categories are 
\sQuote{base}, \sQuote{advanced}, \sQuote{miscellaneous}, \sQuote{scoring}
and \sQuote{usage}.

The query filters are constructed by either \code{\link{filter_per_player}} if
\code{clutch=FALSE} or \code{\link{filter_per_player_clutch}} if
\code{clutch=TRUE}. The list of potential filters are documented in those
functions. The units that the values/data is aggregated into is determined
by the \code{per} filter. The value of \code{clutch} argument is used to
to plays that occured during clutch time.
}
\examples{
\dontrun{
  # Get data to compare clutch vs regular performance.
  ppd <- per_player_agg(filter_per_player(season=2014))
  ppdc <- per_player_agg(filter_per_player_clutch(season=2014))
}

}
\seealso{
\code{\link{filter_per_player}}
  \code{\link{filter_per_player_clutch}}
  \url{http://stats.nba.com/help/glossary}
}

