% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{player_game_logs}
\alias{player_game_logs}
\title{Retrieve player game logs}
\usage{
player_game_logs(league, season, season_type)
}
\arguments{
\item{league}{A character value of league e.g. 'nba', 'd-league'.}

\item{season}{A numeric value of the base season, e.g. 2015 for the 2015-2016
season and so on.}

\item{season_type}{A character value for the type of season, valid types are
\code{regular}, \code{allstar}, \code{playoffs} and \code{preseason}.}
}
\value{
TODO A data.frame with names of current and historical players and the
  fields
  \describe{
     \item{person_id}{integer - ID of the player}
     \item{first_name}{character - the player's first name}
     \item{last_name}{character - the player's last name}
     \item{roster_status}{logical - currently rostered in season selected}
     \item{year_start}{integer - year player entered NBA}
     \item{year_end}{integer - year player exited NBA}
     \item{team_id}{integer - ID of the player's current team (if 
       roster_status is true)}
     \item{team_city}{character - city of current team (if roster_status
       is true)}
     \item{team_name}{character - name of current team (if roster_status is
       true)}
     \item{team_abbr}{character - abbrev. name of current team (if
       roster_status is true)}
     \item{has_played}{logical - did the player record at least one game in
       NBA}
  }
}
\description{
Retrieves all player game logs from stats.nba.com for a given season and
league.
}
\details{
TODO This function returns per-player game data, which is mainly useful for
retrieving \code{game_id} values if per-game shot charts are desired.
}
\examples{
\dontrun{
    df <- get_game_log(league='nba',
                       season=2013,
                       season_type='playoffs')
    tail(df)
}

}
\seealso{
\code{\link{team_game_logs}}
}

