% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{type_converters}
\alias{type_converters}
\title{Named type conversion functions}
\format{An object of class \code{list} of length 12.}
\usage{
type_converters
}
\description{
Descriptiony
}
\keyword{datasets}

